agentms-backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
transformers==4.44.2
pydantic==2.9.2
python-dotenv==1.0.1
torchaudio
torchvision 
torch 
google-generativeai
